requirements.txt
asgiref==3.4.1
certifi==2021.10.8
Django==3.2.9
djangorestframework==3.12.4
interchange==2021.0.4
monotonic==1.6
packaging==21.2
pansi==2020.7.3
Pillow==8.4.0
py2neo==2021.2.3
Pygments==2.10.0
pyparsing==2.4.7
pytz==2021.3
Shapely==1.7.1
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
typish==1.9.3
urllib3==1.26.7